requirements.txt
click==8.1.3
dnspython==2.3.0
email-validator==1.3.1
Flask==2.2.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2-binary==2.9.5
python-decouple==3.7
python-dotenv==0.21.1
SQLAlchemy==2.0.1
typing_extensions==4.4.0
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.12.0
